Refactoring: component analysis functions went into its own module
